Complete Grafix Professional Website - Production Ready
Features:
- Full admin dashboard with authentication
- Product management (CRUD with image uploads)
- Blog system with rich content
- Team/Staff management
- Business settings configuration
- Review moderation system
- Contact form and message inbox
- Newsletter subscriber management
- MySQL database with Flyway migrations
- Spring Security with BCrypt encryption
- Production-ready configuration
- Docker support
- Comprehensive documentation

+++ src/main/resources/db/migration/V2__Add_Reviews_And_Subscribers.sql
@@ -1,6 +1,6 @@
 -- Add Reviews and Subscribers tables
 CREATE TABLE reviews (
-    id INT PRIMARY KEY AUTO_INCREMENT,
+    id BIGINT PRIMARY KEY AUTO_INCREMENT,
     name VARCHAR(100),
     email VARCHAR(100),
     contact VARCHAR(20),
@@ -11,7 +11,7 @@ CREATE TABLE reviews (
 );
 
 CREATE TABLE subscribers (
-    id INT PRIMARY KEY AUTO_INCREMENT,
+    id BIGINT PRIMARY KEY AUTO_INCREMENT,
     email VARCHAR(100) UNIQUE NOT NULL,
     subscribed_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
